Inline opus into rmsnorm.py (drop rmsnorm_opus.py); make opus the default
Merge all opus Python bindings/wrappers into aiter/ops/rmsnorm.py and delete the
separate rmsnorm_opus.py, so rmsnorm.py is the single place for the op. The opus
wrappers (rms_norm_opus, fused_add_rms_norm_opus, rmsnorm2d_fwd_opus,
rmsnorm2d_fwd_with_add_opus, *_dynamicquant_opus, *_smoothquant_opus) are a
complete bf16/fp16 implementation covering plain / fused-add / dynamic+smooth
quant (int8/fp8) / T5, any hidden size -- so the CK (_ck) functions are now a
removable opt-in.

AITER_RMSNORM_BACKEND now defaults to 'opus' (set =ck for the legacy CK path).
With opus default and self-contained, all CK bindings can be deleted later with
no functional loss (gemma_norm / group_size / shuffle_scale keep using the
separate module_rmsnorm_quant, not CK). C++ TU (module_rmsnorm_opus) unchanged.
